Moving placard components to sub folder called UI folder since these really arent related to the other components which are mostly networking/gameplay related.
