+++ Gem/Code/Source/MultiplayerSampleModule.cpp
@@ -11,7 +11,7 @@
 #include <Components/PerfTest/NetworkPrefabSpawnerComponent.h>
 #include <Components/PerfTest/NetworkRandomImpulseComponent.h>
 #include <Components/PerfTest/NetworkTestSpawnerComponent.h>
-#include <Components/UiCanvasDemoPlacardComponent.h>
+#include <Components/UI/UiCanvasDemoPlacardComponent.h>
 #include <Source/AutoGen/AutoComponentTypes.h>
 
 #include "MultiplayerSampleSystemComponent.h"
